my in progress project
Hey chris, so the question i had in my email is: So I was working on my
particle system for my brick breaker game and I tried putting the code
in many different parts of my Brick.h file (all code is going to be
attached in a zip file) and when I put it in draw I noticed after
playing it a bit that it randomly does the particle burst with the
bricks and not all the time like I wanted. here is the code that i have
now and i was wondering what is it that I need to do to get it to where
it bursts all the time the ball hits a brick.
-Miriam